fix: move conn-mode filteration logic from backend-config to eventfilter
